Move invoke tests to subfolder
