+++ dotnet/src/IntegrationTests/Agents/CommonInterfaceConformance/InvokeConformance/AzureAIAgentInvokeTests.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-namespace SemanticKernel.IntegrationTests.Agents.CommonInterfaceConformance;
+namespace SemanticKernel.IntegrationTests.Agents.CommonInterfaceConformance.InvokeConformance;
 
 public class AzureAIAgentInvokeTests() : InvokeTests(() => new AzureAIAgentFixture())
 {
